restfulapi/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
coverage==5.5
distlib==0.3.0
Django==3.1.2
django-cors-headers==3.7.0
django-countries==7.0
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.8.4
flake8-isort==4.0.0
google-api-core==1.23.0
google-api-python-client==1.7.11
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.8
importlib-metadata==1.6.0
isort==5.7.0
mccabe==0.6.1
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pygame==1.9.6
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
testfixtures==6.17.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.0.20
wget==3.2
youtube-dl==2020.1.24
zipp==3.1.0
